Fix the chats list width saving with passcode.
Regression was introduced in 9fabb182f7.

If the passcode is enabled the main widget is created after the main
window was fully initialized, so the first main widget resize already
used the shrink algorithm while creating the third column.

No we shrink the left column only after the main widget was started.