Revert "Fixed adding caption to grouped files."
This reverts commit 351a484264a8aed2569e08b7011a088eaa8d1322.

When sending an album of files each one of them can have its own
caption that will be shown below the file.

A caption for the whole album (in case of media albums) is added
to the first album item. In case of media albums this caption is
displayed under the whole album mosaic. But in case of an album
of files this caption will be displayed under the first file,
between the first and the second file of the album. This is not
what user can expect when he adds a caption for an album.

So we will send it as a separate comment, like it was done before.