Use absolute executable path for dbus service
After the v4.8.7 release, I cannot launch telegram desktop because of the
recent dbus support introduced in
https://github.com/telegramdesktop/tdesktop/commit/9d7544d57cf20b7aea2e2f4a6c3f9b478389a392.
And it was fixed after I changed the value of the Exec field to absolute
path to the telegram-desktop executable.

In dbus specification they don't mention that they will find
executable from PATH variable, so this is a UB.
After this commit, cmake will use the full install path to generate the correct service file
and fix the above issue.

Signed-off-by: Avimitin <dev@avimit.in>