Fix archive open in single-column layout.
Regression was introduced in 83a4cb2bfd.

Before we always force-closed the current filter,
which always force-closed the current opened chat.

Now we don't touch the filter if it is All Chats already,
so that it's scroll state isn't forgotten. In that case
we have to clear the stack in single-column layout explicitly.